#484 Remove unused values in LogviewerReducer and list
