+++ src/js/view/components/helpers/logHelper.js
@@ -25,7 +25,6 @@ export const scrollToBottom = (el, list) => {
 export const fetchNewLinesFromBackendCache = (
   sourcePath,
   nrOfLogLines,
-  feCacheLength,
   indexForNewLines,
   totalLineCountOfFile
 ) => {
@@ -34,11 +33,29 @@ export const fetchNewLinesFromBackendCache = (
     data: {
       sourcePath,
       nrOfLogLines,
-      feCacheLength,
       indexForNewLines,
       totalLineCountOfFile
     }
   };
-  console.log({ argObj });
   sendRequestToBackend(argObj);
+};
+
+export const updateLogViewerCache = cache_size => {
+  const insertRows = (startIndex, newLines) => {
+    const updatedCache = new Array(cache_size).fill('.', 0);
+
+    const fromIndex =
+      startIndex < 0
+        ? 0
+        : newLines.length + startIndex > cache_size
+        ? cache_size - newLines.length
+        : startIndex;
+
+    newLines.forEach((item, i) => {
+      updatedCache[i + fromIndex] = item;
+    });
+
+    return updatedCache;
+  };
+  return { insertRows };
 };
